Fix loading of javascript script language
JavaScriptScriptLanguage depends on nashort which is located in the Java
lib/ext folder, which is loaded by the "Extension ClassLoader". This is
not a parent of the ResourceAwareClassLoader we use. Instead we need to
join the current thread class loader for the extension class loader we
need for loading the Nashorn script engine and the ResourceAwareClassLoader
which we need for Scijava plugins and services.

Signed-off-by: Squareys <Squareys@googlemail.com>